html/browsers/history/the-history-interface/history_replacestate_err.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>history_replaceState SECURITY_ERR</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    try {
      window.history.replaceState(1, document.title, 'http://www.microsoft.com/test.html');
    } catch (e) {
      test(function() {
        assert_equals(e.code, e["SECURITY_ERR"], "SECURITY_ERR is not raised.");
      });
    }
    </script>
  </body>
</html>
